feat: add new route for collector profile and refactor MyCollection component
